+++ app/src/main/java/edu/utep/cs/cs4330/mypricewatcher/ScraperBehavior.java
@@ -18,7 +18,7 @@ public class ScraperBehavior implements PriceFindBehavior {
      * @see Item
      */
     @Override
-    public Long findPrice(Item item) {
+    public Double findPrice(Item item) {
         /* TODO */
         return null;
     }
